# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "10초 가계부 API",
    "description": "앱인토스 미니앱의 백엔드. 인증은 X-Anon-Key 헤더 하나뿐이고 로그인 화면이 없다.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/imports/text": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Analyze Text",
        "operationId": "analyze_text_api_v1_imports_text_post",
        "parameters": [
          {
            "name": "X-Anon-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Anon-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportTextIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatchOut"
                }
              }
            }
          },
          "401": {
            "description": "식별키가 없거나 검증에 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "없거나 내 것이 아님",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "되돌리기 만료·동시 저장",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "하루에 쓸 수 있는 만큼을 넘김",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "검증 서버가 일시적으로 응답하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/{batch_id}/candidates/{candidate_id}": {
      "patch": {
        "tags": [
          "imports"
        ],
        "summary": "Patch Candidate",
        "operationId": "patch_candidate_api_v1_imports__batch_id__candidates__candidate_id__patch",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          },
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          },
          {
            "name": "X-Anon-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Anon-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportCandidatePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatchOut"
                }
              }
            }
          },
          "401": {
            "description": "식별키가 없거나 검증에 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "없거나 내 것이 아님",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "되돌리기 만료·동시 저장",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "하루에 쓸 수 있는 만큼을 넘김",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "검증 서버가 일시적으로 응답하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/{batch_id}/commit": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Commit",
        "operationId": "commit_api_v1_imports__batch_id__commit_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          },
          {
            "name": "X-Anon-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Anon-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportCommitOut"
                }
              }
            }
          },
          "401": {
            "description": "식별키가 없거나 검증에 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "없거나 내 것이 아님",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "되돌리기 만료·동시 저장",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "하루에 쓸 수 있는 만큼을 넘김",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "검증 서버가 일시적으로 응답하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/{batch_id}": {
      "delete": {
        "tags": [
          "imports"
        ],
        "summary": "Destroy",
        "operationId": "destroy_api_v1_imports__batch_id__delete",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          },
          {
            "name": "X-Anon-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Anon-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "식별키가 없거나 검증에 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "없거나 내 것이 아님",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "되돌리기 만료·동시 저장",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "하루에 쓸 수 있는 만큼을 넘김",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "검증 서버가 일시적으로 응답하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant-rules": {
      "get": {
        "tags": [
          "merchant-rules"
        ],
        "summary": "Index",
        "operationId": "index_api_v1_merchant_rules_get",
        "parameters": [
          {
            "name": "X-Anon-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Anon-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantRuleListOut"
                }
              }
            }
          },
          "401": {
            "description": "식별키가 없거나 검증에 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "없거나 내 것이 아님",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "되돌리기 만료·동시 저장",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "하루에 쓸 수 있는 만큼을 넘김",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "검증 서버가 일시적으로 응답하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant-rules/{rule_id}": {
      "delete": {
        "tags": [
          "merchant-rules"
        ],
        "summary": "Destroy",
        "operationId": "destroy_api_v1_merchant_rules__rule_id__delete",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          },
          {
            "name": "X-Anon-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Anon-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "식별키가 없거나 검증에 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "없거나 내 것이 아님",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "되돌리기 만료·동시 저장",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "하루에 쓸 수 있는 만큼을 넘김",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "검증 서버가 일시적으로 응답하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
      "FeedbackKind": {
        "type": "string",
        "enum": [
          "over_budget",
          "pace_warning",
          "large_expense",
          "achievement",
          "on_track",
          "month_fact"
        ],
        "title": "FeedbackKind"
      },
      "FeedbackOut": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/FeedbackKind"
          },
          "remaining_budget": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Budget"
          },
          "daily_allowance": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Allowance"
          },
          "remaining_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Days"
          },
          "over_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Over Amount"
          },
          "over_category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Over Category Id"
          },
          "saved_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved Amount"
          },
          "month_expense": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month Expense"
          },
          "pace_ratio": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pace Ratio"
          },
          "projected_month_end": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Month End"
          },
          "category_spend": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Spend"
          },
          "category_budget_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Budget Amount"
          },
          "large_expense_threshold": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Large Expense Threshold"
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "FeedbackOut",
        "description": "저장 직후 돌려주는 판정 결과.\n\n문장이 아니라 숫자와 종류만 준다. 문장 조립은 화면이 한다."
      },
      "ImportBatchOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source": {
            "$ref": "#/components/schemas/TransactionSource"
          },
          "status": {
            "$ref": "#/components/schemas/ImportBatchStatus"
          },
          "detected_count": {
            "type": "integer",
            "title": "Detected Count"
          },
          "selected_count": {
            "type": "integer",
            "title": "Selected Count"
          },
          "selected_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Selected Total"
          },
          "meta": {
            "$ref": "#/components/schemas/ImportMetaOut"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ImportCandidateOut"
            },
            "type": "array",
            "title": "Candidates"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "status",
          "detected_count",
          "selected_count",
          "selected_total",
          "meta"
        ],
        "title": "ImportBatchOut"
      },
      "ImportBatchStatus": {
        "type": "string",
        "enum": [
          "pending",
          "analyzing",
          "ready",
          "committed",
          "failed"
        ],
        "title": "ImportBatchStatus"
      },
      "ImportCandidateOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "is_low_confidence": {
            "type": "boolean",
            "title": "Is Low Confidence"
          },
          "is_duplicate": {
            "type": "boolean",
            "title": "Is Duplicate"
          },
          "is_selected": {
            "type": "boolean",
            "title": "Is Selected"
          }
        },
        "type": "object",
        "required": [
          "id",
          "occurred_at",
          "amount",
          "type",
          "confidence",
          "is_low_confidence",
          "is_duplicate",
          "is_selected"
        ],
        "title": "ImportCandidateOut"
      },
      "ImportCandidatePatch": {
        "properties": {
          "occurred_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 99999999999999.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransactionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "is_selected": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Selected"
          }
        },
        "type": "object",
        "title": "ImportCandidatePatch",
        "description": "후보 한 줄 고치기. 보낸 항목만 바뀐다.\n\n`merchant` 와 `category_id` 는 명시적으로 null 을 보내면 비운다.\n'안 보냄' 과 'null 로 보냄' 을 가르려고 라우터가 exclude_unset 으로 넘긴다."
      },
      "ImportCommitOut": {
        "properties": {
          "batch": {
            "$ref": "#/components/schemas/ImportBatchOut"
          },
          "created_count": {
            "type": "integer",
            "title": "Created Count"
          },
          "total_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedbackOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetStateOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "batch",
          "created_count",
          "total_amount"
        ],
        "title": "ImportCommitOut",
        "description": "저장 결과. 마지막 한 건 기준으로 지금 돈 상태를 함께 준다."
      },
      "ImportMetaOut": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "is_stub": {
            "type": "boolean",
            "title": "Is Stub"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "is_stub"
        ],
        "title": "ImportMetaOut",
        "description": "이 결과가 어디서 나왔는지. 스텁 결과를 진짜 성공으로 오해하지 않게 한다."
      },
      "ImportTextIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ImportTextIn",
        "description": "줄글 한 덩어리. 여러 건이 들어 있을 수 있다."
      },
      "MerchantRuleListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MerchantRuleOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "MerchantRuleListOut"
      },
...
      "TransactionSource": {
        "type": "string",
        "enum": [
          "keypad",
          "nl",
          "screenshot",
          "receipt",
          "asset_screenshot",
          "no_spend"
        ],
        "title": "TransactionSource",
        "description": "어떤 입력 경로로 들어왔는지. 값을 늘릴 때 고칠 자리는 여기 하나다."
      },
      "TransactionType": {
        "type": "string",
        "enum": [
          "expense",
          "income",
          "transfer",
          "refund"
        ],
        "title": "TransactionType",
        "description": "거래 종류의 유일한 정의. models·schemas·llm 이 전부 이걸 가져다 쓴다."
      },
...
    }
  }
}